DoWork/requirements.txt
asgiref==3.3.4
astor==0.8.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
dataclasses==0.8
decorator==4.4.2
fastapi==0.65.1
gast==0.4.0
gensim==4.0.1
h11==0.12.0
idna==2.10
importlib-metadata==4.5.0
jieba==0.42.1
joblib==1.0.1
numpy==1.19.5
paddlepaddle==2.1.0
pandas==1.1.5
pika==1.2.0
Pillow==8.2.0
protobuf==3.17.1
pydantic==1.8.2
pymongo==3.11.4
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2021.1
redis==3.5.3
requests==2.25.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
smart-open==5.1.0
starlette==0.14.2
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
zipp==3.4.1
